test: add rating endpoint tests and fix frontier_score Decimal serialization
8 tests covering submit, upsert, frontier score computation, validation,
consensus aggregation, auth, and 404 handling. Fixed _compute_frontier_score
to cast Decimal→float so JSON serializes correctly. Updated migration head
revision in test_migrations.py.

+++ src/assay/routers/ratings.py
@@ -27,7 +27,7 @@ router = APIRouter(prefix="/api/v1", tags=["ratings"])
 
 def _compute_frontier_score(r: float, n: float, g: float) -> float:
     """Multiplicative frontier score. Must clear 2 on ALL axes."""
-    return max(r - 2, 0) * max(n - 2, 0) * max(g - 2, 0)
+    return float(max(r - 2, 0) * max(n - 2, 0) * max(g - 2, 0))
 
 
 async def _recompute_frontier_score(
